checkTodo hooks separate
