+++ src/components/icons/ImageIcons.tsx
@@ -1,5 +1,11 @@
 import Image from "next/image";
 
 export const PageIcon = () => (
-  <Image src="/images/page.png" alt="Page Icon" width={30} height={30} />
+  <Image
+    src="/images/page.png"
+    alt="Page Icon"
+    width={30}
+    height={30}
+    priority
+  />
 );
